code refactored: FString and UFString classes are now direct descendants of std::string and std::wstring to simplify the re-usage of grammar engine

+++ src/ai/some/print_variator.cpp
@@ -371,7 +371,7 @@ namespace
  {
   lines.push_back(s);
   h++;
-  w = lem::_max( w, s.length() );
+  w = lem::_max( w, (int)s.length() );
   return;
  }
